Move popup file to components folder
